refactor(iceberg): route registerView through namespaceOperationDispatcher
Per review, registerView now follows the same rule as registerTable: it is part of IcebergNamespaceOperationDispatcher (interface, executor, event dispatcher, hook dispatcher) instead of the view dispatcher chain. IcebergNamespaceOperations no longer injects IcebergViewOperationDispatcher. Moved the import/owner/skip-if-exists hook logic and its tests to IcebergNamespaceHookDispatcher, and the executor test to TestIcebergNamespaceOperationExecutor.
