+++ iceberg/iceberg-rest-server/src/main/java/org/apache/gravitino/iceberg/service/dispatcher/IcebergViewOperationExecutor.java
@@ -24,7 +24,6 @@ import org.apache.gravitino.listener.api.event.IcebergRequestContext;
 import org.apache.iceberg.catalog.Namespace;
 import org.apache.iceberg.catalog.TableIdentifier;
 import org.apache.iceberg.rest.requests.CreateViewRequest;
-import org.apache.iceberg.rest.requests.RegisterViewRequest;
 import org.apache.iceberg.rest.requests.RenameTableRequest;
 import org.apache.iceberg.rest.requests.UpdateTableRequest;
 import org.apache.iceberg.rest.responses.ListTablesResponse;
@@ -44,14 +43,6 @@ public class IcebergViewOperationExecutor implements IcebergViewOperationDispatc
     return icebergCatalogWrapperManager
         .getCatalogWrapper(context.catalogName())
         .createView(namespace, createViewRequest);
-  }
-
-  @Override
-  public LoadViewResponse registerView(
-      IcebergRequestContext context, Namespace namespace, RegisterViewRequest registerViewRequest) {
-    return icebergCatalogWrapperManager
-        .getCatalogWrapper(context.catalogName())
-        .registerView(namespace, registerViewRequest);
   }
 
   @Override
